New files
DBcode.php=is just the code that needs to be pasted in whichever form
that we want it to be in.
DBtablelayout.php= The page where our data is displayed w/our new table.
mailsend.php=Code that I was playing around w/to send email to user and
ourselves. Please feel free to update/add more. HTML part was just
added for testing purposes.